# src/main/resources/static/openapi.yaml
openapi: 3.0.0
info:
  title: Enicilion Platform API
  description: Complete OpenAPI specification for Enicilion Enterprise Java Spring Boot Backend.
  version: 1.0.0
servers:
  - url: http://localhost:4000
    description: Local Development Server
paths:
...
  /api/events/{slug}:
    get:
      summary: Retrieve an event and its active ticket tiers by slug (Public)
      tags:
        - Events
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
          example: motorscape-2026
      responses:
        '200':
          description: Event detail returned successfully
...
  /api/tickets/checkout:
    post:
      summary: Book spectator tickets and create payment intent
      tags:
        - Tickets
      parameters:
        - in: header
          name: Idempotency-Key
          schema:
            type: string
          description: Unique string to prevent duplicate orders
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - items
              properties:
                items:
                  type: array
                  items:
                    type: object
                    required:
                      - eventId
                      - tierId
                      - quantity
                    properties:
                      eventId:
                        type: string
                        format: uuid
                      tierId:
                        type: string
                        format: uuid
                      quantity:
                        type: integer
                        minimum: 1
                        example: 2
                couponCode:
                  type: string
                  example: OFFER10
                clientTotal:
                  type: number
                  example: 1049.00
                guestName:
                  type: string
                  example: Guest User
                guestEmail:
                  type: string
                  format: email
                  example: guest@example.com
                guestPhone:
                  type: string
                  example: "+919876543210"
                idempotencyKey:
                  type: string
      responses:
        '201':
          description: Booking initialized
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      paymentId:
                        type: string
                        format: uuid
                      amount:
                        type: number
                        example: 1049.00
                      currency:
                        type: string
                        example: INR
        '400':
          description: Out of stock, calculation mismatch, or invalid coupon
  /api/tickets/scan:
    post:
      summary: Verify/Check in ticket at the venue gates (Secure Scanner Interface)
      tags:
        - Tickets
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - code
                - password
              properties:
                code:
                  type: string
                  example: MS26-X7G2F9P5-K9WZ
                password:
                  type: string
                  example: enicilion2026
      responses:
        '200':
          description: Verification response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      valid:
                        type: boolean
                        example: true
                      message:
                        type: string
                        example: "✅ Valid! Welcome in."
                      holderName:
                        type: string
                        example: "Jane Doe"
        '401':
          description: Invalid scanner password
  /api/tickets/pdf/{code}:
    get:
      summary: Download PDF tickets by ticket code
      tags:
        - Tickets
      parameters:
        - in: path
          name: code
          required: true
          schema:
            type: string
            example: MS26-X7G2F9P5-K9WZ
      responses:
        '200':
          description: PDF binary file stream
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        '404':
          description: Ticket code not found
  /api/tickets/payment/{paymentId}/pdf:
    get:
      summary: Download all PDF tickets under a payment transaction
      tags:
        - Tickets
      parameters:
        - in: path
          name: paymentId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: PDF binary file stream
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        '400':
          description: Payment is pending/unpaid
        '404':
          description: Payment ID not found
  /api/tickets/{code}/apple-wallet:
    get:
      summary: Generate and download PKPass file for Apple Wallet
      tags:
        - Tickets
      parameters:
        - in: path
          name: code
          required: true
          schema:
            type: string
      responses:
        '200':
          description: PKPass binary stream
          content:
            application/vnd.apple.pkpass:
              schema:
                type: string
                format: binary
  /api/tickets/{code}/google-wallet:
    get:
      summary: Get Google Wallet save-to-pass JWT redirection link
      tags:
        - Tickets
      parameters:
        - in: path
          name: code
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Redirection link
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      saveUrl:
                        type: string
                        example: https://pay.google.com/gp/v/save/eyJhbGci...
  /api/tickets:
    get:
      summary: Retrieve authenticated user's tickets
      tags:
        - Tickets
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User tickets list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        code:
                          type: string
                        tierName:
                          type: string
                        status:
                          type: string
                          enum: [booked, paid, checked_in, cancelled]
...
  /api/admin/events/{id}:
    put:
      summary: Update an existing event (Admin / Staff Only)
      tags:
        - Administration
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - slug
                - location
                - eventDate
              properties:
                name:
                  type: string
                slug:
                  type: string
                description:
                  type: string
                location:
                  type: string
                eventDate:
                  type: string
                  format: date-time
                status:
                  type: string
                  enum: [draft, published, ongoing, completed, cancelled]
                registrationSchema:
                  type: string
                organizerId:
                  type: string
                  format: uuid
      responses:
        '200':
          description: Event updated successfully
    delete:
      summary: Delete an event and all associated tiers (Admin / Staff Only)
      tags:
        - Administration
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Event deleted successfully
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...